# WT-13122: Optimized sesseion's Core time statistics, mongod.log direct locate time-consuming modules, Reduce the difficulty and time of problem analysis