--- a/src/btree/bt_read.c
+++ b/src/btree/bt_read.c
@@ -268,9 +268,11 @@ __wt_page_in_func(WT_SESSION_IMPL *session, WT_REF *ref, uint32_t flags
     WT_REF_STATE current_state;
     WT_TXN *txn;
     uint64_t sleep_usecs, yield_cnt;
+    uint64_t time_start, time_stop;
     int force_attempts;
     bool busy, cache_work, evict_skip, stalled, wont_need;
 
+    time_start = 0;
     btree = S2BT(session);
     txn = session->txn;
 
@@ -316,7 +318,10 @@ read:
              */
             if (!LF_ISSET(WT_READ_IGNORE_CACHE_SIZE))
                 WT_RET(__wt_cache_eviction_check(session, true, txn->mod_count == 0, NULL));
+            time_start = __wt_clock(session);
             WT_RET(__page_read(session, ref, flags));
+            time_stop = __wt_clock(session);
+            WT_STAT_SESSION_INCRV(session, page_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 
             /* We just read a page, don't evict it before we have a chance to use it. */
             evict_skip = true;
@@ -493,6 +498,7 @@ skip_evict:
             return (__wt_illegal_value(session, current_state));
         }
 
+        time_start = __wt_clock(session);
         /*
          * We failed to get the page -- yield before retrying, and if we've yielded enough times,
          * start sleeping so we don't burn CPU to no purpose.
@@ -518,5 +524,8 @@ skip_evict:
         }
         __wt_spin_backoff(&yield_cnt, &sleep_usecs);
         WT_STAT_CONN_INCRV(session, page_sleep, sleep_usecs);
+        time_stop = __wt_clock(session);
+        WT_STAT_SESSION_INCRV(
+          session, ref_locked_and_yield_time, WT_CLOCKDIFF_US(time_stop, time_start));
     }
 }