--- a/src/btree/bt_read.c
+++ b/src/btree/bt_read.c
@@ -521,8 +521,10 @@ skip_evict:
             if (cache_work)
                 continue;
         }
+        time_start = __wt_clock(session);
         __wt_spin_backoff(&yield_cnt, &sleep_usecs);
+        time_stop = __wt_clock(session);
         WT_STAT_CONN_INCRV(session, page_sleep, sleep_usecs);
-        WT_STAT_SESSION_INCRV(session, ref_locked_and_yield_time, sleep_usecs);
+        WT_STAT_SESSION_INCRV(session, ref_locked_and_yield_time, WT_CLOCKDIFF_US(time_stop, time_start));
     }
 }