--- a/src/btree/bt_split.c
+++ b/src/btree/bt_split.c
@@ -2101,7 +2101,7 @@ __wt_split_insert(WT_SESSION_IMPL *session, WT_REF *ref)
      */
     WT_WITH_PAGE_INDEX(session, ret = __split_insert_lock(session, ref));
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, page_split_insert_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, page_split_insert_time, WT_CLOCKDIFF_US(time_stop, time_start));
     WT_STAT_CONN_INCRV(
       session, cache_page_split_insert_time, WT_CLOCKDIFF_MS(time_stop, time_start));
 
@@ -2223,7 +2223,7 @@ __wt_split_multi(WT_SESSION_IMPL *session, WT_REF *ref, int closing)
 
     WT_STAT_CONN_INCRV(
       session, cache_page_split_multi_time, WT_CLOCKDIFF_MS(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, page_split_multi_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, page_split_multi_time, WT_CLOCKDIFF_US(time_stop, time_start));
 
     return (ret);
 }
@@ -2264,7 +2264,7 @@ __wt_split_reverse(WT_SESSION_IMPL *session, WT_REF *ref)
      */
     WT_WITH_PAGE_INDEX(session, ret = __split_reverse(session, ref));
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, page_split_reverse_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, page_split_reverse_time, WT_CLOCKDIFF_US(time_stop, time_start));
     WT_STAT_CONN_INCRV(
       session, cache_page_split_reverse_time, WT_CLOCKDIFF_MS(time_stop, time_start));
 
@@ -2333,7 +2333,7 @@ __wt_split_rewrite(WT_SESSION_IMPL *session, WT_REF *ref, WT_MULTI *multi)
 
     __wt_free(session, new);
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, page_split_rewrite_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, page_split_rewrite_time, WT_CLOCKDIFF_US(time_stop, time_start));
     WT_STAT_CONN_INCRV(
       session, cache_page_split_rewrite_time, WT_CLOCKDIFF_MS(time_stop, time_start));
 
@@ -2342,7 +2342,7 @@ __wt_split_rewrite(WT_SESSION_IMPL *session, WT_REF *ref, WT_MULTI *multi)
 err:
     __split_multi_inmem_fail(session, page, multi, new);
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, page_split_rewrite_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, page_split_rewrite_time, WT_CLOCKDIFF_US(time_stop, time_start));
     WT_STAT_CONN_INCRV(
       session, cache_page_split_rewrite_time, WT_CLOCKDIFF_MS(time_stop, time_start));