--- a/src/cursor/cur_file.c
+++ b/src/cursor/cur_file.c
@@ -199,7 +199,7 @@ err:
     API_RETRYABLE_END(session, ret);
     API_END_RET_STAT(session, ret, cursor_next);
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 }
 
 /*
@@ -233,7 +233,7 @@ __wt_curfile_next_random(WT_CURSOR *cursor)
 err:
     API_END_RET_STAT(session, ret, cursor_next_random);
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 }
 
 /*
@@ -270,7 +270,7 @@ err:
     CURSOR_REPOSITION_END(cursor, session);
     API_END_RET_STAT(session, ret, cursor_prev);
     time_stop = __wt_clock(session);
-    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 }
 
 /*
@@ -343,7 +343,7 @@ err:
     API_END_RET_STAT(session, ret, cursor_search);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opread(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 }
 
 /*
@@ -382,7 +382,7 @@ err:
     API_END_RET_STAT(session, ret, cursor_search_near);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opread(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_read_time, WT_CLOCKDIFF_US(time_stop, time_start));
 }
 
 /*
@@ -424,7 +424,7 @@ err:
     CURSOR_UPDATE_API_END_STAT(session, ret, cursor_insert);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opwrite(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_US(time_stop, time_start));
     return (ret);
 }
 
@@ -493,7 +493,7 @@ err:
     CURSOR_UPDATE_API_END_STAT(session, ret, cursor_modify);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opwrite(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_US(time_stop, time_start));
     return (ret);
 }
 
@@ -528,7 +528,7 @@ err:
     CURSOR_UPDATE_API_END_STAT(session, ret, cursor_update);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opwrite(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_US(time_stop, time_start));
     return (ret);
 }
 
@@ -583,7 +583,7 @@ err:
       session, ret, !positioned || F_ISSET(cursor, WT_CURSTD_KEY_INT), cursor_remove);
     time_stop = __wt_clock(session);
     __wt_stat_usecs_hist_incr_opwrite(session, WT_CLOCKDIFF_US(time_stop, time_start));
-    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_MS(time_stop, time_start));
+    WT_STAT_SESSION_INCRV(session, cursor_write_time, WT_CLOCKDIFF_US(time_stop, time_start));
     return (ret);
 }