… dist/stat_data.py
… src/include/serial_inline.h
… src/include/stat.h
… src/include/wiredtiger.in
… src/log/log.c
… src/reconcile/rec_write.c
… src/support/stat.c