Optimized sesseion's Core time statistics, mongod.log direct locate time-consuming modules, Reduce the difficulty and time of problem analysis.

Finally, I will submit another PR to MongoDB to comply with the session statistics added by WT, so that we can see the corresponding mongod.log log as follows(The logs record the modules in which the user thread consumes more than 1ms during WT execution):
{"t":{"$date":"2024-06-01T15:15:36.782+08:00"},"s":"I"....................."storage":{"total write times": "1111ms"; Submodule time consuming：{"page lock wait":"100ms", "ref lock sleep":"50ms", "reconcile time":"11ms", "split multi time": "11ms", "bytesWrite":9451235,"read time":150592,......, "log write":"33"ms}};{"total read times": "1111ms"; Submodule time consuming: {"page lock wait":"100ms", "ref lock sleep":"50ms", "reconcile time":"11ms", "split multi time": "11ms", "bytesRead":9451235,"read time":150592,......}},"remote":"10.111.0.39:40870","protocol":"op_msg","durationMillis":10580}} 


Recently, I have applied this function to our company's MongoDB test environment (users will do pressure test and Functional test in the test environment before they really go online). Judging from the test experience of nearly two weeks, the effect is quite good. From the slow log, we can visually locate whether the problem is caused by mongo server or wt. if jitter caused by wt,  it Can visually locate the jitter caused by which small wt module.

 

At present, the sample number of our test environment is relatively small, and it is expected that more potential jitter problems of WT will be found in the future large-scale online use. We expect to find all module problems that may cause jitter through this function, and finally solve these jitter problems.

 

Sorry, this PR has not been evaluated by you yet. I am worried that I will forget my design ideas after too long, so I will submit the code first.

If you assess that the PR doesn't make sense to you, we can just shut it down.