@korteland 2024-06-24
The discussion about whether we want to take this change needs to happen on the Jira ticket first, so I'm closing this out until that happens.